se crea las rutas chatsRouter y userRouter
se crean las rutas y se mueve la logica de programacion a chatsRouter
